Remove 'bionic' builders from nacl, chromium and naclports masters
The bionic NaCl toolchain effort is no longer needed
not that we have a working arm/glibc toolchain.

For naclports the builder is transitioning to be a
toolchain (pnacl-on-pnacl) builder.  For the other
waterfalls we no longer require these VMs and they
should be returned to the free pool (vm154-m3, vm319-m3)

Review URL: https://codereview.chromium.org/1489723002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@297750 0039d316-1c4b-4281-b951-d872f2087c98