When parsing values from chartjson, try replacing - with @@.
Background: Since https://codereview.chromium.org/1472973004/, the dashboard
will return a metric string with of the form "interaction-chart/trace" or
"interaction-chart/interaction-chart" if the dashboard guesses that the test
path includes an interaction record label name.

However, in chartjson the interaction record label and trace name may be
separated with @@ rather than -. This CL changes the parsing of results to
try replacing the "-" with "@@" if no results could be extracted at first.

This CL is based on https://codereview.chromium.org/1487163002/, so it contains
all of the changes there. That CL contains only refactoring changes, so those
changes can be reviewed separately.

BUG=564353

Review URL: https://codereview.chromium.org/1492073002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@297803 0039d316-1c4b-4281-b951-d872f2087c98