WebRTC: Add recipes to chromium recipe module.
This is a step towards making WebRTC use the chromium recipe
for all our Chromium builders. No builders are yet switched over to use
these recipes yet.

These changes are a subset of https://codereview.chromium.org/1371293003/
a variety of changes to make the expectations as similar as possible
to the existing recipe.

BUG=497757, 538259

Review URL: https://codereview.chromium.org/1406253003

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@297876 0039d316-1c4b-4281-b951-d872f2087c98