WebRTC: Add media_unittests to our Chromium bots with webcams.
Actually only the bots in chromium.webrtc have physical webcams, but
we're planning on getting webcams for the other ones as well.
In case of no camera, the tests are passing after printing 
"No camera available. Exiting test.", which is why I also configured
the tests to print the stdio.
The tests needs to be executed sequentially to avoid failing due to
device access.

BUG=137260

Review URL: https://codereview.chromium.org/1517503003

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@297938 0039d316-1c4b-4281-b951-d872f2087c98