WebRTC: Remove webrtc.DEPS from Chromium builders that don't need it.
During recent flakiness in the runhooks step caused by the large number of
things we're pulling down and building as part of the DEPS file in
https://chromium.googlesource.com/chromium/deps/webrtc/webrtc.DEPS/+/master/DEPS
it makes sense to remove it from the builders and the GN bots, since
they don't use it. The Android testers also don't need these resources.

This should speed up cycle time a bit.

Review URL: https://codereview.chromium.org/1517663002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@297946 0039d316-1c4b-4281-b951-d872f2087c98