WebRTC: Add remoting_unittests to our Chromium bots.
It makes sense to run this test to detect breakages early on (mainly for
chromium.webrtc.fyi but the test is so quick so it doesn't hurt having it
in chromium.webrtc as well).

This depends on https://codereview.chromium.org/1512143003 being
landed first.

Review URL: https://codereview.chromium.org/1516823002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@298002 0039d316-1c4b-4281-b951-d872f2087c98