Let ClangToTMac and ClangToTMacAsan not auto_reboot after builds.
The same machine is used as builder and tester, and if the builder bot
is configured to autoreboot after a build, it'll reboot the slave right
after it started doing the tester work, resulting in the tester stage
always being interrupted and purple.

Keep the tester on auto_reboot, which will hopefully be enough to keep
the purpleness during builds away (which was why we made these bots
auto_reboot in the first place).

BUG=495186

Review URL: https://codereview.chromium.org/1524223002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@298022 0039d316-1c4b-4281-b951-d872f2087c98