Add bot configurations for OPENSSL_SMALL.
Chromium and Android build with OPENSSL_SMALL defined, so we should have bot
coverage here. Since the default for a standalone build is still the non-SMALL
build, I've kept the existing Windows and Mac builders as-is. With 3-ish
developers and 1 minute cycle times, we've got plenty of spare capacity.

BUG=571243

Review URL: https://codereview.chromium.org/1543693002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@298089 0039d316-1c4b-4281-b951-d872f2087c98