buildbucket: set client_operation_id when triggering
If a build was committed to the datastore on buildbucket, but something
happened on the way back to the client, the client will retry the
request and created a duplicate build. Unless client_operation_id by the
client, which allows the server to associated the second request with
the previous one and return the build created by the first request.

R=sergiyb@chromium.org
BUG=570660

Review URL: https://codereview.chromium.org/1531113004

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@298091 0039d316-1c4b-4281-b951-d872f2087c98