Handle missing metric correctly even when the test doesn't fail.
The missing metric case may or may not cause a failure on the test job.
In case it doesn't fail but still returns an empty list of values, we
handle it similarly to the failure case.

Note that in the return code mode, we don't expect values in the list
and in that case we don't consider this a failure.

BUG=563483
R=prasadv,chrisphan

Review URL: https://codereview.chromium.org/1561433002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@298128 0039d316-1c4b-4281-b951-d872f2087c98