WebRTC: Enable OpenH264 for WebRTC standalone bots.
hbos@ in the WebRTC team is working on enabling OpenH264 as an H.264
encoder in Chrome as part of http://crbug.com/468365. This CL sets the
necessary GYP variables to turn it on in the build for standalone WebRTC
bots and the Chromium bots we control.

NOTICE: Due to licensing and patent reasons, OpenH264 cannot be enabled
by default for end-users. Going forward, the plan is that it will only be
enabled on WebRTC controlled bots and the Chromium official builders.

BUG=576201

Review URL: https://codereview.chromium.org/1577773002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@298217 0039d316-1c4b-4281-b951-d872f2087c98