V8 Buildbot: Fix auto-bisect for slim swarming testers.
This uses the gitiles api instead of local git calls for
resolving the change range. We cannot use git on slim
swarming testers as they don't have a checkout.

Logging the changes in the UI is removed now as it is
completely redundant to the json output of the previous
step.

BUG=535160

Review URL: https://codereview.chromium.org/1594973007

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@298318 0039d316-1c4b-4281-b951-d872f2087c98