[WebView] Fix resource size test on webview bot.
Fixed the paths for the .so libraries for the resource size
test. The paths changed when switching from a builder/tester
to tester bot.

Additionally, added a test in example.py for the resource_size
function and made optional arguments to the resouce_size.py script
optional in the resource_size function.

TBR=phajdan.jr@chromium.org
BUG=

Review URL: https://codereview.chromium.org/1640123006

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@298458 0039d316-1c4b-4281-b951-d872f2087c98