Stop setting werror='' on Clang ToT bots.
On the Mac bots, it has no effects.  Hence, the Win, Mac, and iOS bots ToT bots
currently build with warnings-as-errors and the Linux and Android bots don't.
If a new warning breaks the Mac and Windows bots but not the Linux bots, I'm
always confused for ten minutes, so make the bots behave the same in this
regard.

The Mac/Win behavior feels more useful -- come clang roll, we need to figure
out the warnings anyways, and it's easy to just disable a warning on the ToT
bots if necessary.  (And often, even fixing isn't that hard.)

BUG=none
TBR=phajdan

Review URL: https://codereview.chromium.org/1650253003

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@298497 0039d316-1c4b-4281-b951-d872f2087c98