Configures the codesearch buildbot to allow absolute file paths.
This is because Codesearch seems to require absolute file paths to
function as expected.

In conjunction with codereview.chromium.org/1643523004, this should allow codesearch to
continue working, while still allowing all other buildbots and users
to gain the benefits of not requiring absolute paths.

BUG=580103

Review URL: https://codereview.chromium.org/1642193002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@298498 0039d316-1c4b-4281-b951-d872f2087c98