WebRTC: Build Chromium with codecs
In order to build and test the OpenH264 codec, we need
to pass additional GYP_DEFINES variables to our bots.
These are already used by several bots in Chromium, but 
mainly by the GPU team.
OpenH264 is not yet built when they're defined, but the
plan is to make it so soon, so we want to ensure we're
building with it since Chromium tests marked as MANUAL_
are going to be added, which only run in these 
WebRTC-controlled waterfalls.

The GN bots are handled by the MB config in
https://codereview.chromium.org/1654953004/

BUG=576201

Review URL: https://codereview.chromium.org/1658183002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@298518 0039d316-1c4b-4281-b951-d872f2087c98