WebRTC: Add iOS bots running tests with the simulator
Add recipes and two experimental bots in client.webrtc.fyi
to try this out.

This CL adds support for a 'use_analyze' variable to be specified
in the JSON bot configs in case analyze is not be used.

This depends on buildbot configurations that are added in
https://codereview.webrtc.org/1660053002/

BUG=chromium:498746

Review URL: https://codereview.chromium.org/1659163003

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@298548 0039d316-1c4b-4281-b951-d872f2087c98