Specify win_optional_gpu_tests_rel tryserver using src-side JSON files.
The associated chromium.gpu waterfall bots don't actually exist, but
have entries in the src-side JSON files. This trybot (and others which
are forthcoming) runs additional tests that are otherwise run only on
the chromium.gpu.fyi waterfall.

After this CL, the GPU recipe is unreferenced. It will be deleted in a
subsequent CL.

This CL introduces a couple of TODOs for code which loses coverage once
the GPU tests stop being added manually to at least one bot. These will
be fixed in subsequent CLs.

BUG=542370,583702

Review URL: https://codereview.chromium.org/1669743002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@298597 0039d316-1c4b-4281-b951-d872f2087c98