Only generate flutter docs on linux
This at least reduces the chance of racing rsyncs.
To actually eliminate them we need to either use a separate
builder/recipe which only does the docs and give it a pool
of slaves of size=1, or we need to use a lock on the bucket.

R=abarth@chromium.org

Review URL: https://codereview.chromium.org/1684983002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@298696 0039d316-1c4b-4281-b951-d872f2087c98