[Android] Fix capitalization of WebView for CTS bot.
The WebView bots have inconsistant capitalization. Some are named
WebView and some are named Webview. Fixing this capitalization to
fix the CTS bot not being triggered properly. In the future, I'd
like to change all the capitalization for all bots to be the same.

BUG=
TBR=phajdan.jr@chromium.org

Review URL: https://codereview.chromium.org/1689233002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@298743 0039d316-1c4b-4281-b951-d872f2087c98