Add chromium_clang_upload bot for Linux.
Right now, it checks out Chromium + the deps for Android and calls tools/clang/scripts/package.py with no parameters.

The next step will be to allow package.py to upload the binaries to Google Cloud Storage, if --upload-binaries is specified, and then make the bot to pass this flag.

BUG=578306

Review URL: https://codereview.chromium.org/1688043002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@298751 0039d316-1c4b-4281-b951-d872f2087c98