[Findit] Use analyze to skip unaffected tests in recipe findit/chromium/test.
This is mainly to speed up culprit finding by Findit: if a test target is not affected by a commit (eg: change to telemetry code only won't affect browser_tests), we could skip running the tests in the test target against the unrelated commit.

This could also help a little bit when a test case is flaky: re-running it against an unrelated commit might still result in a test failure, and thus lead to false positive in culprit finding.

BUG=584488

Review URL: https://codereview.chromium.org/1694733002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@298782 0039d316-1c4b-4281-b951-d872f2087c98