art: Fix other builders to fail on any error.
The current behavior is such that it will run all the steps, and cause them to warn on failure, but not fail the overall build when the steps error out. This changes the behavior to be that it runs all the steps, no matter if a step fails, but if any step fails, then it fails the whole build.

See https://build.chromium.org/p/client.art/builders/aosp-builder/builds/5392 for an example of the unwanted behavior, and https://build.chromium.org/p/client.art/builders/aosp-builder/builds/5393 for an example of the desired behavior.

TBR=dnj@chromium.org

BUG=

Review URL: https://codereview.chromium.org/1697263005

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@298803 0039d316-1c4b-4281-b951-d872f2087c98