V8: Pass archive location to fuzzer.
Instead of scraping the fuzz output for the archive name,
this will pass the name to the fuzz harness. This avoids
consuming and adding back the output to the UI and makes
the archive location easy to manipulate.

This makes sure the archives are deleted after fuzzing.

This also cleans up a few unnecessary lines.

Depends on https://codereview.chromium.org/1706203002/

BUG=chromium:535160

Review URL: https://codereview.chromium.org/1711733002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@298850 0039d316-1c4b-4281-b951-d872f2087c98