Add try_recipe to try execute other recipes as tryjobs.
The idea is that if you have a CL which modifies or creates a new recipe,
you may want to test for real, as opposed to just expectations. The
try_recipe is exactly for that.

This CL also adds a tool to indent steps produced by calling the recipe
being tried, so as to improve the rendering of the result by buildbot.

TBR=machenbach@chromium.org

Review URL: https://codereview.chromium.org/1717473002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@298876 0039d316-1c4b-4281-b951-d872f2087c98