WebRTC: Switch all iOS bots to new recipe.
A new recipe is based on the ios recipe module and keeps most of
the build configuration in JSON files in the WebRTC checkout
(webrtc/build/ios directory).

These bots have been staged in the client.webrtc.fyi waterfall, with the
exception of the iOS GN bots, which are now being staged.

Recipes will be removed in a separate CL after the masters have been
restarted (to avoid breaking them in the transition).

BUG=588590
TBR=machenbach@chromium.org

Review URL: https://codereview.chromium.org/1719853002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@298902 0039d316-1c4b-4281-b951-d872f2087c98