Flip linux_chromium_clobber_rel_ng to mirror Linux x64 on p/chromium.
Previously it was "mirroring" a non-existing "Linux Builder (clobber)"
bot on the chromium.fyi waterfall. There was no need to do that
since that is what the "Linux x64" bot is doing on the chromium waterfall.

A nice side effect of this is that the linux_chromium_clobber_rel_ng
bot will also switch to using MB (correctly) and GN.

R=phajdan.jr@chromium.org, jam@chromium.org
BUG=588018

Review URL: https://codereview.chromium.org/1716093002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@298906 0039d316-1c4b-4281-b951-d872f2087c98