Fix how clobber is used in the recipe converter
Buildbot sets "clobber" to an empty string to indicate that the build is
a clobber build. Before this change, the converter assumed that the
clobber property being set to a non-empty string indicated a clobber
build. This CL fixes the converter to use "in" instead of "get".

BUG=480029
R=iannucci,dnj

Review URL: https://codereview.chromium.org/1696313002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@298913 0039d316-1c4b-4281-b951-d872f2087c98