iOS: Fix condition for use_analyze in ios recipe module.
Turns out the default value (enabled) is True, but when
use_analyze is set to 'false' in the JSON, the value becomes
a string 'false'. Fix this by changing the condition check.

This caused the WebRTC trybots to always use analyze, which
is not supported (fails).

TBR=smut@google.com
BUG=588590

Review URL: https://codereview.chromium.org/1723033003

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@298919 0039d316-1c4b-4281-b951-d872f2087c98