WebRTC: Remove poller for webrtc-samples repo.
Having multiple GittilesPollers produces blamelists
with hashes from different repos, but more importantly perf
dashboard links that are not valid, since the from-hash can be
from the Chromium repo and the to-hash from the webrtc-samples repo.

Since breakages by changes in the webrtc-samples are rare, we should
just remove the poller of that repo for simplicity.
What has changed reasonably recently is that we download a prebuilt
AppRTC which is pinned to specific revision and rolled manually, so
if there are changes in AppRTC that are breaking Chromium, we still
wouldn't detect them anymore.

This solves one of the problems in crbug.com/491520 but more work is
needed.

TBR=phoglund@webrtc.org
BUG=491520

Review URL: https://codereview.chromium.org/1754243003

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@299053 0039d316-1c4b-4281-b951-d872f2087c98