Add support for building and uploading Android x86-64 artifacts
I used the storage name android-x64 to match linux-x64,
unclear if that's correct.

Depends on https://github.com/flutter/engine/pull/2455.

This is also duplicating some of the uploads from the arm
version, which is silly.  The jar and keystore should be the same.

It's possible we could fix the GN files to know how to build
Android x86, Android x86-64, Android ARM and Android ARM64
in one build directory.  Also not clear if we need all those.

Currently we build Android ARM and this CL adds Android x86-64.

R=abarth@chromium.org

Review URL: https://codereview.chromium.org/1764893002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@299112 0039d316-1c4b-4281-b951-d872f2087c98