Removing build timeouts.
Some bisect jobs are failing to find a culprit because they time out
waiting for builds. Sheriffs seem to trigger bisects in bursts, which
swamps the builders.

We decided that since we can detect failed builds, there is no sense in
timing out waiting for a build to complete.

R=dtu,prasadv
BUG=593462

Review URL: https://codereview.chromium.org/1784593002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@299191 0039d316-1c4b-4281-b951-d872f2087c98