Use closing_optional for Cronet bots in gatekeeper.json
gatekeeper.json was modified in crrev.com/1724533003 to
always email cronet team on cronet bots failures. However,
that CL doesn't quite work, because that CL didn't not
specify which step to be notified on. This CL uses 
closing_optional with * to match all steps to be notified
on, if any one of them fails.

Review URL: https://codereview.chromium.org/1749173002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@299230 0039d316-1c4b-4281-b951-d872f2087c98