Use buildername rather than slavename for "bot" with perf dashboard.
Rationale: This is what is used most commonly for the "bot" names on the perf dashboard, and the builder name usually gives a clearer idea of the platform (rather than the slave name, which often doesn't tell you the platform and can change if buildslaves are changed).

The libvpx/android_unittest recipe uses the perf_dashboard recipe module; this CL changes that recipe to explicitly pass the slave name since that is what it has done in the past. Note that the expectations don't change for that recipe.

BUG=593431

Review URL: https://codereview.chromium.org/1782723002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@299231 0039d316-1c4b-4281-b951-d872f2087c98