Fix zip_build.py converter and chromium.fyi recipe
The Chromium Builder was incorrectly passing a null value for the
build_archive_url to its triggered builders on chromium.fyi due to the
recipe not setting the build property that zip_build.py hitherto set.

This Cl fixes that problem in the recipe, as well as changing the recipe
converter to fix this in all subsequent automatic conversions as well.

BUG=480029
R=dnj,iannucci

Review URL: https://codereview.chromium.org/1788433002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@299237 0039d316-1c4b-4281-b951-d872f2087c98