Make all bots on chromium.webkit tree-closers.
Previously we had different rules for the Leak bots and the non-Oilpan bots, but now they should be treated just like the regular bots. We also had entries for the deps bots, which were removed as part of the merge of Blink into Chromium.

R=haraken@chromium.org
BUG=

Review URL: https://codereview.chromium.org/1781153002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@299243 0039d316-1c4b-4281-b951-d872f2087c98