Add 'Linux LTO Perf' buildbot (recipes part).
The src part of the change: https://codereview.chromium.org/1728193006/

This buildbot will emulate regular Linux Perf builders:
https://build.chromium.org/p/chromium.perf/builders/Linux%20Builder

The difference is that the build will use Link-Time Optimization.
This is much slower, and before we could launch, we need to know
how many slaves we should have to build for every single
Chromium revision.

This buildbot will not archive the output or upload it to GCS.

BUG=586664

Review URL: https://codereview.chromium.org/1808283002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@299366 0039d316-1c4b-4281-b951-d872f2087c98