Match Skia Swarming buildslaves to number of swarming bots
At the moment, with all buildslaves able to trigger any task but with
fewer Windows and Mac Swarming bots, the buildbots tend to group up on
the Windows tasks and get stuck. As a temporary measure, let's group the
buildslaves to match the number of swarming bots for each OS, so that
buildslaves never have to wait for swarming bots.

TBR=KevinL
BUG=skia:4763

Review URL: https://codereview.chromium.org/1814973003

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@299382 0039d316-1c4b-4281-b951-d872f2087c98