Remove remaining Linux 32-bit builders that we want to drop.
This removes the following bots:
 - 'Google Chrome Linux' on the chromium.chrome waterfall
 - 'Linux Trusty (32)' and 'Linux Trusty (32)(dbg) on the
   chromium.fyi waterfall
 - 'linux_chromium_archive_rel_ng' on the tryserver.chromium.linux
   waterfall.

We will keep the existing 'Chromium Linux32 Goma Canary (clobber)'
bot on the chromium.fyi waterfall for the moment, and keep
'Linux Builder (dbg)(32)' and 'Linux Tests (dbg)(1)(32)' on
the chromium.linux waterfall (and matching trybots) indefinitely.

We should probably *add* optional release 32-bit builders, and re-image
the existing 32-bit builders, so that they are using a normal 64-bit
image and just using the 32-bit sysroots, but that'll happen in a
follow-up CL once we can shuffle bot pools a bit.

R=ukai@chromium.org, phajdan.jr@chromium.org
BUG=590004

Review URL: https://codereview.chromium.org/1808423002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@299392 0039d316-1c4b-4281-b951-d872f2087c98