goma_utils: use time.mktime to calcualte timestamp from epoch.
In US Pacific, timezone could be either of PDT and PST.
If we did not specify timezone to python, local timezone would be used.
If we calculate timedelta from epoch, when it is in PDT, timedelta
could be off by an hour.
Then, ts_mon would get future start time.

BUG=595260

Review URL: https://codereview.chromium.org/1814803004

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@299394 0039d316-1c4b-4281-b951-d872f2087c98