Add log and link aliasing to BuildBot.
This adds support for log/url aliases to BuildBot, and augments the
annotation engine to extract aliases from specially-formatted STEP_LINK
commands.

BuildBot will now associate and render a log/url's alias alongside it in
the build view.

The BuildBot annotation parser will identify STEP_LINK of the form
"alias-->orig", and add it as an alias for "orig" called "alias" instead
of as a separate URL link.

The template and BuildBot changes are intended to be backwards-compatible
with currently-running BuildBot instances to allow for master restarts
to pick this up gradually. A master that hasn't restarted will render
alias annotations as just more links (e.g., "alias-->orig"), should
aliases be emitted to that master before it has been restarted.

BUG=chromium:595946
TEST=local
  - First up a custom waterfall / recipe engine, ran, looks decent.

Review URL: https://codereview.chromium.org/1812163002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@299396 0039d316-1c4b-4281-b951-d872f2087c98