Don't add 'SKIPPED' tests to the list of failures.
The android instrumentation tests emit SKIPPED for tests that aren't available on the current platform (for example, some tests are only available on L+). The current GTest parser puts SKIPPED into the list of failures, which seems wrong.

The upside of this change is more test frameworks can use the GTest framework instead of their own framework, and we can reduce the complexity in the buildsystem.

BUG=525873

Review URL: https://codereview.chromium.org/1849253003

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@299699 0039d316-1c4b-4281-b951-d872f2087c98