Give the clang upload bots shorter names.
Currently, git fails to delete files on these bots because path names
are too long (on Windows), and the bot / checkout name is part of the
path.  Our other bots don't have this problem, so try giving the bot
a shorter name (and rename the non-Win bots for consistency).

BUG=600465

Review URL: https://codereview.chromium.org/1858343002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@299707 0039d316-1c4b-4281-b951-d872f2087c98