Auto-reboot the chromium.gpu and chromium.gpu.fyi bots.
Auto-reboots were disabled some time ago in the spirit of efficiency,
but this caused problems like lingering processes on Windows which were
severely detrimental.

The only four slaves that do not auto-reboot are Linux slaves which use
the "subdir" property to run multiple slaves on the same physical host
and can not be auto-rebooted.

BUG=601640

Review URL: https://codereview.chromium.org/1866403003

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@299786 0039d316-1c4b-4281-b951-d872f2087c98