Flip WebKit Mac 10.9 to be a builder/tester.
For some reason the Mac10.9 bot is generating different
layout test baselines than then mac_blink_rel trybots.
It appears that the system images are the same and the
compile flags are the same.

The next working hypothesis is that there's some difference
resulting from building on 10.11 and testing on 10.9 versus
building on 10.9 and testing directly; the most likely culprit
seems to be Xcode, which has different versions on the two bots.

TBR=iannucci@chromium.org
BUG=601166

Review URL: https://codereview.chromium.org/1881323002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@299879 0039d316-1c4b-4281-b951-d872f2087c98