Equalizing samples when comparing middle revision against good and bad.
When examining a revision in the bisect process, we want to decide which
of the halves of the range between 'good' and 'bad' revisions to either
side of the revison being tested is more likely to contain the source of
the regression. By taking into account samples of the same size for
either end of the range, we hope to reduce any bias that the sample size
may impart on the hypothesis test used to perform the comparison.

R=dtu
BUG=597998

Review URL: https://codereview.chromium.org/1911583003

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@300060 0039d316-1c4b-4281-b951-d872f2087c98