Flip builders on chromium.memory.fyi to MB.
This flips all of the builders on the chromium.memory.fyi
waterfall to MB (except for "Chromium Linux Builder (valgrind)",
which was already on it). Some of these are not yet on recipes,
but we're flipping them anyway to avoid blocking the GN migration
on the recipe migration.

R=brettw@chromium.org, aneeshm@chromium.org
BUG=481692

Review URL: https://codereview.chromium.org/1878503002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@300063 0039d316-1c4b-4281-b951-d872f2087c98