Send mac_chromium_rel_ng's tryjobs to new AMD Retina machines.
There are more of these now in the Swarming pool than the old NVIDIA
Retina machines.

Unfortunately it's not easy to send a tryjob through for this change,
but the associated waterfall bots are running well. Once this change has
landed and stuck, all of the old NVIDIA Retina bots will be
decommissioned.

BUG=572793

Review URL: https://codereview.chromium.org/1902273006

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@300064 0039d316-1c4b-4281-b951-d872f2087c98