Remove 'Blink Linux LSAN MSAN' bot from chromium.fyi.
We don't have anyone available to work on migrating this
to the main waterfall, recipes, and GN, unfortunately, so
I'm going to remove this.

Bug 361678 is still open, so hopefully someone can pick it
up in the future and bring the bot online using a recipe, which
should be simple enough given that we have ASAN and MSAN
bots already.

TBR=thestig@chromium.org
BUG=605732, 361678

Review URL: https://codereview.chromium.org/1904363002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@300099 0039d316-1c4b-4281-b951-d872f2087c98