Switch chromium_utils.RunCommand's stdout reader thread to read one line at a time, instead of one byte for layout tests on Windows trybots.
This was changed in http://src.chromium.org/viewvc/chrome?view=rev&revision=57562 to fix missing output when a test hangs. Since then, most if not all the test runners have their own timeouts. Even if a test hangs, missing the last line is not critical now if it fixes this bug. If it doesn't, I'll revert.

BUG=522396

Review URL: https://codereview.chromium.org/1919483003

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@300124 0039d316-1c4b-4281-b951-d872f2087c98