Make path to tests absolute if possible on bisect.
Switching from naked commands to runtest.py had the side effect of
changing the working directory where the command is run from. This broke
bisect. To fix this, and to prevent changes in the future from breaking
this again, we will try to make this an absolute path.

R=prasadv
BUG=606281

Review URL: https://codereview.chromium.org/1920023003

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@300173 0039d316-1c4b-4281-b951-d872f2087c98