This moves all of the remaining 32-bit linux builders onto 64-bit machines.
The last phase of cleaning up the Linux 32 configs is to replace
the builders and testers with normal 64-bit machines; we no
longer wish to support the weird '64-bit kernel with 32-bit userspace'
configs that we had before.

The waterfall bots are repurposed valgrind machines. The tryservers
are moved into the existing optional pool of machines, since they
won't be used often.

R=thestig@chromium.org
BUG=600903, 600904, 590004

Review URL: https://codereview.chromium.org/1917883003

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@300213 0039d316-1c4b-4281-b951-d872f2087c98