Adding test generators to the Nexus 9 bot.
Previously, the tests were not running on the Nexus 9 bot.
We believe this is because there were no test generators.
This should enable a handful of test generators, including telemetry ones.
But since there are no telemetry tests on that bot, those should effectively
be a noop.

BUG=606160
TBR=kbr@chromium.org

Review URL: https://codereview.chromium.org/1924653002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@300214 0039d316-1c4b-4281-b951-d872f2087c98