RunCommand: Reworking the process and thread waiting code.
Instead of using .wait() which will block us forever in the kernel, we poll
things and sleep for short periods. While /in theory/ the polling *shouldn't*
be any different to doing a .wait() it can help nudge the threads out of the
blocked state.

Doing this also makes the code more reliable and resilient to when processes
and threads finish in random order and unreliable delivery of notifications.

This is a speculative patch based on a hunch. The correct solution to this
problem is to rewrite this function or remove runtests.py all together.

BUG=522396
R=jam,dpranke

Review-Url: https://codereview.chromium.org/1911853003

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@300261 0039d316-1c4b-4281-b951-d872f2087c98