Call run-webkit-tests directly on windows.
In yet another attempt to eliminate layers of wrappers to see
if it helps with the Windows hangs, this CL stops calling
layout_test_wrapper.py and instead just calls run-webkit-tests
directly. This is a generally good change and one that has been
put off for a long time, so even if this doesn't fix the hang
we should probably leave it in, though we'd need to add the
xvfb functionality somewhere to replace it on Linux.

TBR=tansell@chromium.org, ojan@chromium.org
BUG=522396

Review-Url: https://codereview.chromium.org/1934893002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@300385 0039d316-1c4b-4281-b951-d872f2087c98