Revert all of the run-webkit-test wrapper changes from this weekend.
We probably want to keep some variant of this code in, but such a
decision should be properly reviewed (and tested across platforms).
For now, this CL backs out the changes I landed over the previous
few days (except for the src-side change to run-webkit-tests that
added the --json-test-results flag; that change was simple, harmless,
and will be needed if we end up wanting these changes).

TBR=tansell@chromium.org, jam@chromium.org, ojan@chromium.org
BUG=522396

This reverts commit e81f2296279d56342be5be503bc1322ce1778c78.

Revert "Call run-webkit-tests directly on windows."

This reverts commit fea188005db5f9d19ac15381721a0fa75209e31d.

Revert "Fix command line arg passed to run-webkit-tests"

This reverts commit d1d2df9ecbda3c8d66bdfd2b7003b646ebbf805c.

Revert "Tweak invocation of run-webkit-tests to fix bugs."

This reverts commit 6964e8331716a59206aaaed1b67686cd1fc0fad4.

BUG=

Review-Url: https://codereview.chromium.org/1936943002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@300389 0039d316-1c4b-4281-b951-d872f2087c98