BuildBot: Add anchors for unbound aliases.
Currently, aliases only attach themselves to logs/links that reference
them. If those logs/links disappear, the alias is not shown. This change
makes it so that aliases without log/link anchors will generate link
anchors. This solves the problem of log aliase disappearing when the
logs are pruned from the master.

BUG=chromium:602384
TEST=local
  - Created aliased builds, deleted logs, still render.

R=estaab@chromium.org, iannucci@chromium.org, martiniss@chromium.org, sergiyb@chromium.org

Review-Url: https://codereview.chromium.org/1885033002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@300424 0039d316-1c4b-4281-b951-d872f2087c98