Fail build if triggering fails
Introduce the concept of "fatal error" to AnnotatedCommand. In case of
fatal errors, mark entire build as EXCEPTION and put the details in the
fatal_errors log of the preamble.

If 'critical' is True or absent in a triggering spec and triggering
fails on the master side, it is a fatal error.

R=vadimsh@chromium.org
BUG=609166

Review-Url: https://codereview.chromium.org/1949753003

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@300505 0039d316-1c4b-4281-b951-d872f2087c98