Remove the run-webkit-test wrappers from the chromium recipe.
This change re-lands the changes that were reverted in r300389,
so that the chromium recipe calls run-webkit-tests directly,
rather than calling through the runtest.py and layout_test_wrapper
wrappers. This will remove some opportunities for hangs in
python subprocess code, and if this change sticks on windows
we can probably change it a bit and use this on mac and windows
as well, getting rid of the wrapper altogether.

R=tansell@chromium.org, iannucci@chromium.org
BUG=522396

Review-Url: https://codereview.chromium.org/1965233002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@300545 0039d316-1c4b-4281-b951-d872f2087c98