Always kill off subprocesses in twistd within 30 seconds of their termination.
Previously twisted would wait for the combination of the process being lost,
plus the closure of stdout, stderr and stdin. This change makes twisted start a
30 second timer from the moment the reactor is signalled that the process has
died. If 30 seconds passes and the std* handles are still not closed, we
consider the process gone, and move on with whatever was waiting for the
termination of the process.

R=dnj@chromium.org, dpranke@chromium.org, estaab@chromium.org
BUG=522396

Review-Url: https://codereview.chromium.org/1970563003

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@300550 0039d316-1c4b-4281-b951-d872f2087c98