Revert of Add hack to make webrtc paches work on chromium trybots. (patchset #1 id:1 of https://codereview.chromium.org/1983873003/ )
Reason for revert:
The problem seems to lie elsewhere. 
I'm also not sure this is what we want: the problem with the WebRTC patches is that the patch contains paths starting with webrtc/, so src/third_party as cwd for the patch (git apply) command is the right thing (until crbug.com/611808 is fixed).

Original issue's description:
> Add hack to make webrtc paches work on chromium trybots.
> 
> BUG=438952
> TBR=kjellander@chromium.org
> 
> Committed: http://src.chromium.org/viewvc/chrome?view=rev&revision=300605

TBR=tandrii@chromium.org
# Skipping CQ checks because original CL landed less than 1 days ago.
NOPRESUBMIT=true
NOTREECHECKS=true
NOTRY=true
BUG=438952

Review-Url: https://codereview.chromium.org/1987603002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@300614 0039d316-1c4b-4281-b951-d872f2087c98