recipes: Switch from 3 ways for requesting clobber builds to 1
Some bots applied the clobber config, some set force_clobber when
calling compile(), and some set clobber: true on the bot. Move to the
first, and remove compile()'s force_clobber parameter and make it
look only at the first signal.

The motivation is that once this is done in a unified way, clobber can
be moved to a separate step that runs before runhooks and
generate_buidl_files instead of during compile – then clobbering doesn't
have to be careful to not delete ninja files, isolate files, and other things
that are generated by runhooks and generate_build_files.

No intended behavior change.

BUG=574557

Review-Url: https://codereview.chromium.org/2033513002